+++ src/ploomber/cli/task.py
@@ -1,6 +1,8 @@
 from ploomber.cli.parsers import _custom_command, CustomParser
 
 
+# TODO: we are just smoke testing this, we need to improve the tests
+# (check the appropriate functions are called)
 def main():
     parser = CustomParser(description='Get task information')
     with parser:
@@ -22,6 +24,10 @@ def main():
                             '-st',
                             help='Get task status',
                             action='store_true')
+        parser.add_argument('--on-finish',
+                            '-of',
+                            help='Only execute on_finish hook',
+                            action='store_true')
     dag, args = _custom_command(parser)
 
     dag.render(force=args.force)
@@ -33,8 +39,11 @@ def main():
     if args.status:
         print(task.status())
 
-    # task if built by default, but when --source or --status are passed,
+    if args.on_finish:
+        task._run_on_finish()
+
+    # task if build by default, but when --source or --status are passed,
     # the --build flag is required
-    no_flags = not any((args.build, args.status, args.source))
+    no_flags = not any((args.build, args.status, args.source, args.on_finish))
     if no_flags or args.build:
         task.build(force=args.force)