+++ src/ploomber/spec/TaskSpec.py
@@ -85,6 +85,10 @@ def source_for_task_class(source_str, task_class, project_root, lazy_import,
     else:
         path = Path(source_str)
 
+        # NOTE: there is some inconsistent behavior here. project_root
+        # will be none if DAGSpec was initialized with a dictionary, hence
+        # this won't resolve to absolute paths - this is a bit confusing.
+        # maybe always convert to absolute?
         if project_root and not path.is_absolute() and make_absolute:
             return Path(project_root, source_str)
         else:
@@ -122,6 +126,7 @@ class TaskSpec(MutableMapping):
         self.data = data
         self.meta = meta
         self.project_root = project_root
+        self.lazy_import = lazy_import
 
         self.validate()
 
@@ -198,6 +203,15 @@ class TaskSpec(MutableMapping):
         on_finish = task_dict.pop('on_finish', None)
         on_render = task_dict.pop('on_render', None)
         on_failure = task_dict.pop('on_failure', None)
+
+        # edge case: if using lazy_import, we should not check if the kernel
+        # is installed. this is used when exporting to Argo/Airflow using
+        # soopervisor, since the exporting process should not require to have
+        # the ir kernel installed. The same applies when Airflow has to convert
+        # the DAG, the Airflow environment shouldn't require the ir kernel
+        if (class_ == tasks.NotebookRunner and self.lazy_import
+                and 'check_if_kernel_installed' not in task_dict):
+            task_dict['check_if_kernel_installed'] = False
 
         task = class_(source=source,
                       product=product,