+++ tests/spec/test_DAGSpec.py
@@ -424,6 +424,22 @@ def test_pipeline_r(tmp_pipeline_r):
     dag.build()
 
 
+def test_initialize_with_lazy_import_with_missing_kernel(
+        tmp_pipeline_r, monkeypatch):
+    Path('output').mkdir()
+
+    with open('pipeline.yaml') as f:
+        dag_spec = yaml.load(f, Loader=yaml.SafeLoader)
+
+    def no_kernel(name):
+        raise ValueError
+
+    monkeypatch.setattr(jupyter_client.kernelspec, 'get_kernel_spec',
+                        no_kernel)
+
+    assert DAGSpec(dag_spec, lazy_import=True).to_dag()
+
+
 @pytest.mark.parametrize('raw', [[{
     'source': 'load.py'
 }], {
@@ -666,7 +682,26 @@ def test_import_tasks_from(tmp_nbs):
     spec = DAGSpec(spec_d)
 
     spec.to_dag().render()
-    assert 'extra_task.py' in [str(t['source']) for t in spec['tasks']]
+    assert str(Path('extra_task.py').resolve()) in [
+        str(t['source']) for t in spec['tasks']
+    ]
+
+
+def test_import_tasks_from_does_not_change_dotted_paths(tmp_nbs):
+    some_tasks = [{
+        'source': 'extra_task.py',
+        'product': 'extra.ipynb'
+    }, {
+        'source': 'test_pkg.touch_root',
+        'product': 'some_file.csv'
+    }]
+    Path('some_tasks.yaml').write_text(yaml.dump(some_tasks))
+
+    spec_d = yaml.safe_load(Path('pipeline.yaml').read_text())
+    spec_d['meta']['import_tasks_from'] = 'some_tasks.yaml'
+
+    spec = DAGSpec(spec_d, lazy_import=True)
+    assert 'test_pkg.touch_root' in [t['source'] for t in spec['tasks']]
 
 
 def test_import_tasks_from_with_non_empty_env(tmp_nbs):
@@ -691,4 +726,89 @@ def test_import_tasks_from_with_non_empty_env(tmp_nbs):
     dag = spec.to_dag()
     dag.render()
     assert dag['extra_task'].params['some_param'] == 'some_value'
-    assert 'extra_task.py' in [str(t['source']) for t in spec['tasks']]
+    assert str(Path('extra_task.py').resolve()) in [
+        str(t['source']) for t in spec['tasks']
+    ]
+
+
+def test_import_tasks_from_loads_relative_to_pipeline_spec(tmp_nbs):
+    some_tasks = [{'source': 'extra_task.py', 'product': 'extra.ipynb'}]
+    Path('some_tasks.yaml').write_text(yaml.dump(some_tasks))
+    Path('extra_task.py').write_text("""
+# + tags=["parameters"]
+# -
+""")
+
+    spec_d = yaml.safe_load(Path('pipeline.yaml').read_text())
+    spec_d['meta']['import_tasks_from'] = 'some_tasks.yaml'
+
+    Path('pipeline.yaml').write_text(yaml.dump(spec_d))
+
+    # move to another dir to make sure we can still load the spec
+    Path('subdir').mkdir()
+    os.chdir('subdir')
+
+    spec = DAGSpec('../pipeline.yaml')
+    dag = spec.to_dag()
+    dag.render()
+
+    assert spec['meta']['import_tasks_from'] == str(
+        Path('..', 'some_tasks.yaml').resolve())
+    assert str(Path('..', 'extra_task.py').resolve()) in [
+        str(t['source']) for t in spec['tasks']
+    ]
+
+
+def test_import_tasks_from_keeps_value_if_already_absolute(tmp_nbs, tmp_path):
+    tasks_yaml = (tmp_path / 'some_tasks.yaml').resolve()
+    path_to_script = (tmp_path / 'extra_task.py').resolve()
+
+    some_tasks = [{'source': str(path_to_script), 'product': 'extra.ipynb'}]
+    tasks_yaml.write_text(yaml.dump(some_tasks))
+    path_to_script.write_text("""
+# + tags=["parameters"]
+# -
+""")
+
+    spec_d = yaml.safe_load(Path('pipeline.yaml').read_text())
+    # set an absolute path
+    spec_d['meta']['import_tasks_from'] = str(tasks_yaml)
+    Path('pipeline.yaml').write_text(yaml.dump(spec_d))
+
+    spec = DAGSpec('pipeline.yaml')
+    dag = spec.to_dag()
+    dag.render()
+
+    # value should be the same because it was absolute
+    assert spec['meta']['import_tasks_from'] == str(tasks_yaml)
+    assert str(path_to_script) in [str(t['source']) for t in spec['tasks']]
+
+
+def test_import_tasks_from_paths_are_relative_to_the_yaml_spec(
+        tmp_nbs, tmp_path):
+    tasks_yaml = tmp_path / 'some_tasks.yaml'
+
+    # source is a relative path
+    some_tasks = [{'source': 'extra_task.py', 'product': 'extra.ipynb'}]
+    tasks_yaml.write_text(yaml.dump(some_tasks))
+    #  write the source code in the same folder as some_tasks.yaml
+    Path(tmp_path, 'extra_task.py').write_text("""
+# + tags=["parameters"]
+# -
+""")
+
+    spec_d = yaml.safe_load(Path('pipeline.yaml').read_text())
+
+    # set an absolute path
+    spec_d['meta']['import_tasks_from'] = str(tasks_yaml.resolve())
+    Path('pipeline.yaml').write_text(yaml.dump(spec_d))
+
+    spec = DAGSpec('pipeline.yaml')
+    dag = spec.to_dag()
+    dag.render()
+
+    # paths must be interpreted as relative to tasks.yaml, not to the
+    # current working directory
+    assert str(Path(tmp_path, 'extra_task.py').resolve()) in [
+        str(t['source']) for t in spec['tasks']
+    ]